requirements.txt
faiss-cpu==1.7.4
numpy==2.2.6
pandas==2.3.3
sentence_transformers==5.1.0
torch==2.9.1
